Convert several functions to pass Property(RowSet, Row, Value)_r objects instead of S(RowSet, Row, PropValue)
I propogated the change reasonably far, until I ran into other functions
that are still using the S variants. I then found the cast operator and
cast a PropertyRowSet to a SRowSet case fairly close to where it was returned.

Instead of all propogating all the type changes it might make sense
to do the cast at the mapi function that needs the Property*_r object.